Warning fixes

+++ options/posix/include/sys/socket.h
@@ -58,7 +58,7 @@ struct cmsghdr {
 // Returns a pointer to the next header or nullptr if there is none.
 #define CMSG_NXTHDR(m, c) \
 	((c)->cmsg_len < sizeof(struct cmsghdr) || \
-		sizeof(struct cmsghdr) + CMSG_ALIGN((c)->cmsg_len) \
+		(ptrdiff_t)(sizeof(struct cmsghdr) + CMSG_ALIGN((c)->cmsg_len)) \
 			>= __MLIBC_MHDR_LIMIT(m) - (char *)(c) \
 	? (struct cmsghdr *)0 : (struct cmsghdr *)__MLIBC_CMSG_NEXT(c))
